test(agents): extract shared MCP harness fixtures to test-support module
The harness test file grew past the max-lines lint cap once the approval
gating and OAuth bootstrap provenance tests landed. Move the shared
makeRuntime/makeConnectRuntime fixtures into a test-support module
(mirroring mcp-ui-resource.test-support.js) instead of adding a new test
file, which would push the agents-root core test shard past its 720-root
boundary.
