+++ src/agents/agent-bundle-mcp-harness.test-support.ts
@@ -1,0 +1,102 @@
+import { createRequesterMcpConnect } from "./agent-bundle-mcp-requester-connect.js";
+/** Shared runtime fixtures for the agent bundle MCP harness materializer tests. */
+import type { SessionMcpRuntime } from "./agent-bundle-mcp-types.js";
+
+export function makeRuntime(params: {
+  sessionId: string;
+  requesterSenderId: string;
+}): SessionMcpRuntime {
+  const serverName = "user-mail";
+  const catalog = {
+    version: 1,
+    generatedAt: 0,
+    servers: {
+      [serverName]: {
+        serverName,
+        launchSummary: serverName,
+        toolCount: 1,
+      },
+    },
+    tools: [
+      {
+        serverName,
+        safeServerName: serverName,
+        toolName: "inbox",
+        description: "read inbox",
+        inputSchema: { type: "object", properties: {} },
+        fallbackDescription: "read inbox",
+      },
+    ],
+  };
+  let lastUsedAt = Date.now();
+  let activeLeases = 0;
+  return {
+    sessionId: params.sessionId,
+    workspaceDir: "/workspace",
+    configFingerprint: "fp",
+    requesterScope: { requesterSenderId: params.requesterSenderId },
+    createdAt: Date.now(),
+    get lastUsedAt() {
+      return lastUsedAt;
+    },
+    get activeLeases() {
+      return activeLeases;
+    },
+    acquireLease: () => {
+      activeLeases += 1;
+      let released = false;
+      return () => {
+        if (released) {
+          return;
+        }
+        released = true;
+        activeLeases -= 1;
+      };
+    },
+    markUsed: () => {
+      lastUsedAt = Date.now();
+    },
+    peekCatalog: () => catalog,
+    getCatalog: async () => catalog,
+    callTool: async (_server, toolName) => ({
+      content: [
+        {
+          type: "text",
+          text: `live:${toolName}:${params.requesterSenderId}`,
+        },
+      ],
+      isError: false,
+    }),
+    dispose: async () => {},
+  };
+}
+
+export async function makeConnectRuntime(params: {
+  sessionId: string;
+  requesterSenderId: string;
+  publicOrigin?: string;
+}): Promise<SessionMcpRuntime> {
+  const runtime = makeRuntime(params);
+  const catalog = { version: 1, generatedAt: 0, servers: {}, tools: [] };
+  runtime.peekCatalog = () => catalog;
+  runtime.getCatalog = async () => catalog;
+  runtime.requesterConnect = await createRequesterMcpConnect({
+    serverNames: new Set(["calendar"]),
+    mcpServers: {
+      calendar: {
+        url: "https://mcp.example/rpc",
+        auth: "oauth",
+        oauth: { identity: "per-requester" },
+      },
+    },
+    safeServerNamesByServer: new Map([["calendar", "calendar"]]),
+    requesterScope: {
+      requesterSenderId: params.requesterSenderId,
+      messageChannel: "telegram",
+      agentAccountId: "bot",
+    },
+    cfg: params.publicOrigin ? { gateway: { publicOrigin: params.publicOrigin } } : undefined,
+    configFingerprint: "connect-fingerprint",
+  });
+  return runtime;
+}
